Update đăng nhập, đăng kí, biểu đồ

+++ src/main/java/com/example/demo/model/ChiTieu.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,7 +19,7 @@ public class ChiTieu {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private Double soTien;
+    private BigDecimal soTien;
 
     private String moTa;
